Rename commit "undo" terminology to uncommit
Standardize terminology from "undo" to "uncommit" across API, types, and
client bindings to make intent clearer and reflect that changes remain
as uncommitted workspace modifications. Updated type names, field names,
function signatures, and tests to use UncommitResult / uncommittedIds,
and adjusted imports/usages and generated SDK types accordingly.
